Add files via upload
Basically, this project was to increase the data set for machine learning tasks artifically. We used the dataset of kaggle, plant disease, we picked up the potato as plant and its three diseases for which we have to increase dataset. We totally have just 2600 images 224,224,3 for this task. Although the quality of the images was not perfect because of smaller dimensions and dataset, but we get some usable results.